move event args out of SeekerApplication
